+++ tests/Fixtures/App/Document/AllStrategiesDocument.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Cmf\Bundle\SeoBundle\Tests\Resources\Document;
+namespace Symfony\Cmf\Bundle\SeoBundle\Tests\Fixtures\App\Document;
 
 use Symfony\Cmf\Bundle\RoutingBundle\Doctrine\Phpcr\Route;
 use Symfony\Cmf\Bundle\SeoBundle\Doctrine\Phpcr\SeoMetadata;