+++ tests/Fixtures/App/Document/SeoAwareContent.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Cmf\Bundle\SeoBundle\Tests\Resources\Document;
+namespace Symfony\Cmf\Bundle\SeoBundle\Tests\Fixtures\App\Document;
 
 use Doctrine\ODM\PHPCR\Mapping\Annotations as PHPCRODM;
 use Symfony\Cmf\Bundle\SeoBundle\Model\SeoMetadataInterface;