+++ tests/Fixtures/App/config/bundles.php
@@ -1,0 +1,17 @@
+<?php
+
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2017 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+return [
+    Sonata\SeoBundle\SonataSeoBundle::class => ['all' => true],
+    Symfony\Cmf\Bundle\SeoBundle\CmfSeoBundle::class => ['all' => true],
+    Symfony\Cmf\Bundle\RoutingBundle\CmfRoutingBundle::class => ['all' => true],
+    Symfony\Cmf\Bundle\CoreBundle\CmfCoreBundle::class => ['all' => true],
+];