+++ tests/Unit/DependencyInjection/Compiler/RegisterSuggestionProviderPassTest.php
@@ -43,14 +43,15 @@ class RegisterSuggestionProviderPassTest extends AbstractCompilerPassTestCase
         $this->setDefinition('matcher.with_group', $matcherServiceWithGroup);
 
         $matcherPresentationService = new Definition();
+        $matcherPresentationService->setArguments([1, 1, 2, 3, []]);
         $this->setDefinition('cmf_seo.error.suggestion_provider.controller', $matcherPresentationService);
 
         $this->compile();
 
-        $this->assertContainerBuilderHasServiceDefinitionWithMethodCall(
+        $this->assertContainerBuilderHasServiceDefinitionWithArgument(
             'cmf_seo.error.suggestion_provider.controller',
-            'addSuggestionProvider',
-            [new Reference('matcher.with_group'), 'some-group']
+            4,
+            [['provider' => new Reference('matcher.with_group'), 'group' => 'some-group']]
         );
     }
 
@@ -64,14 +65,15 @@ class RegisterSuggestionProviderPassTest extends AbstractCompilerPassTestCase
         $this->setDefinition('matcher.without_group', $matcherServiceWithOutGroup);
 
         $matcherPresentationService = new Definition();
+        $matcherPresentationService->setArguments([1, 1, 2, 3, []]);
         $this->setDefinition('cmf_seo.error.suggestion_provider.controller', $matcherPresentationService);
 
         $this->compile();
 
-        $this->assertContainerBuilderHasServiceDefinitionWithMethodCall(
+        $this->assertContainerBuilderHasServiceDefinitionWithArgument(
             'cmf_seo.error.suggestion_provider.controller',
-            'addSuggestionProvider',
-            [new Reference('matcher.without_group'), 'default']
+            4,
+            [['provider' => new Reference('matcher.without_group'), 'group' => 'default']]
         );
     }
 }